requirements.txt
fastapi[standard]
langchain
openai
pydotenv
langchain_community
langchain_groq